+++ backend-t-shirt/src/main/java/dk/ss/backendtshirt/tshirt/repository/GiftConfigRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GiftConfigRepository extends JpaRepository<GiftConfig,Integer> {
+public interface GiftConfigRepository extends JpaRepository<GiftConfig,Long> {
     // Vi antager, at der kun er én aktiv config, eller vi henter den med ID 1
     GiftConfig findFirstByActive(int active);
